Initial commit of the video player countdown app.
The purpose of this app is to randomly play videos in a given directory until a set time, and to display a countdown timer as that time approaches.  This could be useful for displaying ASL songs (or some other type of media) before the start of a meeting (or after a meeting but before the meeting for service).